requirements.txt
beautifulsoup4==4.12.2
Flask==3.0.0
google_api_python_client==2.108.0
google_auth_oauthlib==1.1.0
python_dateutil==2.8.2
Requests==2.31.0
Flask-Cors==4.0.0